07/30/2026

+++ server/src/services/reports.ts
@@ -1,22 +1,19 @@
-import express from "express";
+import express from 'express'
 
-import { prisma } from "../server";
-import fs from "fs";
-import path from "path";
+import { prisma } from '../server'
+import fs from 'fs'
+import path from 'path'
 
-const REPORTS_BASE_URL = "http://localhost:3000/reports";
+const REPORTS_BASE_URL = 'http://localhost:3000/reports'
 
-async function changeReport(
-  request: express.Request,
-  response: express.Response,
-) {
-  const bearer = request.headers.authorization;
-  const token = bearer?.split(" ")[1];
+async function changeReport(request: express.Request, response: express.Response): Promise<any> {
+  const bearer = request.headers.authorization
+  const token = bearer?.split(' ')[1]
 
   if (!token) {
     return response.status(400).json({
-      message: "Invalid credentials.",
-    });
+      message: 'Invalid credentials.',
+    })
   }
 
   const user = await prisma.user.findUnique({
@@ -27,21 +24,21 @@ async function changeReport(
       id: true,
       report: true,
     },
-  });
+  })
 
   if (!user) {
     return response.status(404).json({
-      message: "Account not exists.",
-    });
+      message: 'Account not exists.',
+    })
   }
 
-  const oldReport = user.report?.reportUrl;
-  const fileName = request.file?.filename;
+  const oldReport = user.report?.reportUrl
+  const fileName = request.file?.filename
 
   if (!fileName) {
     return response.status(400).json({
-      message: "File sent not found.",
-    });
+      message: 'File sent not found.',
+    })
   }
 
   try {
@@ -60,42 +57,39 @@ async function changeReport(
           },
         },
       },
-    });
+    })
 
-    if (oldReport)
-      fs.unlinkSync(
-        path.resolve(__dirname, "..", "..", "database", "reports", oldReport),
-      );
+    if (oldReport) fs.unlinkSync(path.resolve(__dirname, '..', '..', 'database', 'reports', oldReport))
   } catch (error) {
     return response.status(400).json({
-      message: "Error while tried to change report!",
-    });
+      message: 'Error while tried to change report!',
+    })
   }
 
   const report = await prisma.report.findUnique({
     where: {
       userId: user.id,
     },
-  });
+  })
 
   if (!user) {
     return response.status(404).json({
-      message: "Account not exists.",
-    });
+      message: 'Account not exists.',
+    })
   }
 
   if (!report) {
     return response.status(400).json({
-      message: "Could not change report.",
-    });
+      message: 'Could not change report.',
+    })
   }
 
-  report.reportUrl = `${REPORTS_BASE_URL}/${report?.reportUrl}`;
+  report.reportUrl = `${REPORTS_BASE_URL}/${report?.reportUrl}`
 
   response.status(200).json({
-    message: "Report changed.",
+    message: 'Report changed.',
     report,
-  });
+  })
 }
 
-export { changeReport };
+export { changeReport }
